requirements.txt
annotated-types==0.6.0
anyio==3.7.1
beautifulsoup4==4.12.2
click==8.1.7
fastapi==0.103.2
h11==0.14.0
idna==3.4
lxml==4.9.3
pydantic==2.4.2
pydantic_core==2.10.1
python-docx==1.0.1
python-multipart==0.0.6
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
